+++ src/bl64-check.bash
@@ -650,7 +650,7 @@ function bl64_check_module() {
   local setup_status=''
 
   bl64_check_parameter 'module' &&
-    bl64_check_module_imported "$module" ||
+    bl64_lib_module_imported "$module" ||
     return $?
 
   eval setup_status="\$$module"
@@ -659,34 +659,6 @@ function bl64_check_module() {
     return $BL64_LIB_ERROR_MODULE_SETUP_MISSING
   fi
 
-  return 0
-}
-
-#######################################
-# Check that the module is imported
-#
-# * Used for the modular version of bashlib64 to ensure dependant modules are loaded (sourced)
-# * A module is considered imported if the associated shell environment variable BL64_XXX_MODULE is defined
-# * This check will not verify if the module was also initialized. Use the function 'bl64_check_module' instead
-#
-# Arguments:
-#   $1: module id (eg: BL64_XXXX_MODULE)
-# Outputs:
-#   STDOUT: none
-#   STDERR: message
-# Returns:
-#   0: check ok
-#   BL64_LIB_ERROR_MODULE_NOT_IMPORTED
-#######################################
-function bl64_check_module_imported() {
-  bl64_dbg_lib_check_enabled && bl64_dbg_lib_show_function "$@"
-  local module="${1:-}"
-  bl64_check_parameter 'module' || return $?
-
-  if [[ ! -v "$module" ]]; then
-    bl64_msg_show_error "${_BL64_CHECK_TXT_MODULE_SET} (${_BL64_CHECK_TXT_MODULE}: ${module} ${BL64_MSG_COSMETIC_PIPE} ${_BL64_CHECK_TXT_FUNCTION}: ${FUNCNAME[1]:-NONE}@${BASH_LINENO[1]:-NONE}.${FUNCNAME[2]:-NONE}@${BASH_LINENO[2]:-NONE})"
-    return $BL64_LIB_ERROR_MODULE_NOT_IMPORTED
-  fi
   return 0
 }