+++ src/bl64-pkg-setup.bash
@@ -21,11 +21,11 @@ function bl64_pkg_setup() {
   bl64_dbg_lib_show_function
 
   # shellcheck disable=SC2249
-  bl64_check_module_imported 'BL64_DBG_MODULE' &&
-    bl64_check_module_imported 'BL64_CHECK_MODULE' &&
-    bl64_check_module_imported 'BL64_MSG_MODULE' &&
-    bl64_check_module_imported 'BL64_FS_MODULE' &&
-    bl64_check_module_imported 'BL64_RXTX_MODULE' &&
+  bl64_lib_module_imported 'BL64_DBG_MODULE' &&
+    bl64_lib_module_imported 'BL64_CHECK_MODULE' &&
+    bl64_lib_module_imported 'BL64_MSG_MODULE' &&
+    bl64_lib_module_imported 'BL64_FS_MODULE' &&
+    bl64_lib_module_imported 'BL64_RXTX_MODULE' &&
     _bl64_pkg_set_command &&
     _bl64_pkg_set_runtime &&
     _bl64_pkg_set_options &&